requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.20.26
botocore==1.23.54
Django==4.0.1
djangorestframework==3.13.1
jmespath==0.10.0
mysqlclient==2.1.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
